Remove unused components, and chore fix UI
