--- src/components/Splash.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import splash from '../assets/images/splash.svg'
-
-const Splash: React.FC = () => {
-  return (
-    <>
-      <img src={splash} alt="Splash image"></img>
-    </>
-  )
-}
-
-export default Splash
